MOVE App.js to components folder
